+++ syndicate/core/conf/validator.py
@@ -13,6 +13,7 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 """
+import datetime
 import os
 
 MIN_BUCKET_NAME_LEN = 3
@@ -37,6 +38,16 @@ DEPLOY_TARGET_BUCKET_CFG = 'deploy_target_bucket'
 PROJECTS_MAPPING_CFG = 'build_projects_mapping'
 RESOURCES_SUFFIX_CFG = 'resources_suffix'
 RESOURCES_PREFIX_CFG = 'resources_prefix'
+
+USE_TEMP_CREDS_CFG = 'use_temp_creds'
+SERIAL_NUMBER_CFG = 'serial_number'
+TEMP_AWS_ACCESS_KEY_ID_CFG = 'temp_aws_access_key_id'
+TEMP_AWS_SECRET_ACCESS_KEY_CFG = 'temp_aws_secret_access_key'
+TEMP_AWS_SESSION_TOKEN_CFG = 'temp_aws_session_token'
+EXPIRATION_CFG = 'expiration'
+ACCESS_ROLE_CFG = 'access_role'
+
+TAGS_CFG = 'tags'
 
 PYTHON_LANGUAGE_NAME = 'python'
 NODEJS_LANGUAGE_NAME = 'nodejs'
@@ -80,7 +91,39 @@ class ConfigValidator:
                 VALIDATOR: self._validate_resources_prefix_suffix},
             RESOURCES_SUFFIX_CFG: {
                 REQUIRED: False,
-                VALIDATOR: self._validate_resources_prefix_suffix}
+                VALIDATOR: self._validate_resources_prefix_suffix},
+            USE_TEMP_CREDS_CFG: {
+                REQUIRED: False,
+                VALIDATOR: self._validate_use_temp_creds
+            },
+            SERIAL_NUMBER_CFG: {
+                REQUIRED: False,
+                VALIDATOR: self._validate_serial_number
+            },
+            TEMP_AWS_SECRET_ACCESS_KEY_CFG: {
+                REQUIRED: False,
+                VALIDATOR: self._validate_aws_access_key
+            },
+            TEMP_AWS_ACCESS_KEY_ID_CFG: {
+                REQUIRED: False,
+                VALIDATOR: self._validate_aws_secret_access_key
+            },
+            TEMP_AWS_SESSION_TOKEN_CFG: {
+                REQUIRED: False,
+                VALIDATOR: self._validate_aws_session_token
+            },
+            EXPIRATION_CFG: {
+                REQUIRED: False,
+                VALIDATOR: self._validate_expiration
+            },
+            ACCESS_ROLE_CFG: {
+                REQUIRED: False,
+                VALIDATOR: self._validate_access_role
+            },
+            TAGS_CFG: {
+                REQUIRED: False,
+                VALIDATOR: self._validate_tags
+            }
         }
 
     def validate(self):
@@ -96,7 +139,6 @@ class ConfigValidator:
                 if not value:
                     error_messages[key] = REQUIRED_PARAM_ERROR.format(key)
                     continue
-
             validator_func = validation_rules.get(VALIDATOR)
             validation_errors = validator_func(key, value)
             if validation_errors:
@@ -191,6 +233,64 @@ class ConfigValidator:
         if str_error:
             return [str_error]
 
+    def _validate_aws_session_token(self, key, value):
+        str_error = self._assert_value_is_str(key=key,
+                                              value=value)
+        if str_error:
+            return [str_error]
+
+    def _validate_use_temp_creds(self, key, value):
+        bool_error = self._assert_value_is_bool(
+            key=key, value=value
+        )
+        if bool_error:
+            return [bool_error]
+
+    def _validate_serial_number(self, key, value):
+        str_error = self._assert_value_is_str(key=key,
+                                              value=value)
+        if str_error:
+            return [str_error]
+
+    def _validate_access_role(self, key, value):
+        str_error = self._assert_value_is_str(key=key,
+                                              value=value)
+        if str_error:
+            return [str_error]
+
+    def _validate_tags(self, key, value):
+        errors = []
+        if not value:
+            return errors
+        if not isinstance(value, dict):
+            errors.append(f'\'{key}\' param must be a dictionary but '
+                          f'not a \'{type(value).__name__}\'')
+            return errors
+        if len(value) > 50:
+            errors.append(f'Each resource can have up to 50 user created tags.'
+                          f' You have specified: {len(value)}')
+        for tag_name, tag_value in value.items():
+            if tag_name.startswith('aws:'):
+                errors.append(f'\'{tag_name}\': you can\'t create, edit or '
+                              f'delete a tag that begins with the \'aws:\' '
+                              f'prefix.')
+                if not 1 <= len(tag_name) <= 128:
+                    errors.append(f'\'{tag_name}\': the tag key must be a '
+                                  f'minimum of 1 and a maximum of 128 Unicode '
+                                  f'characters')
+                if not 0 <= len(tag_value) <= 256:
+                    errors.append(f'\'{tag_value}\': the tag value must be a '
+                                  f'minimum of 0 and a maximum of 256 Unicode '
+                                  f'characters')
+        return errors
+
+
+    @staticmethod
+    def _validate_expiration(key, value):
+        if not isinstance(value, datetime.datetime):
+            return [f'\'{key}\' must be a valid ISO 8601 format string']
+        return []
+
     @staticmethod
     def _validate_resources_prefix_suffix(key, value):
         str_error = ConfigValidator._assert_value_is_str(key=key, value=value)
@@ -210,3 +310,8 @@ class ConfigValidator:
     def _assert_value_is_str(key, value):
         if type(value) is not str:
             return f'{key} must be type of string'
+
+    @staticmethod
+    def _assert_value_is_bool(key, value):
+        if type(value) is not bool:
+            return f'{key} must be type of bool'