+++ syndicate/core/handlers.py
@@ -45,23 +45,31 @@ from syndicate.core.build.warmup_processor import (process_deploy_resources,
 from syndicate.core.conf.validator import (JAVA_LANGUAGE_NAME,
                                            PYTHON_LANGUAGE_NAME,
                                            NODEJS_LANGUAGE_NAME)
-from syndicate.core.decorators import check_deploy_name_for_duplicates
+from syndicate.core.decorators import (check_deploy_name_for_duplicates,
+                                       check_deploy_bucket_exists)
 from syndicate.core.groups.generate import (generate,
                                             GENERATE_PROJECT_COMMAND_NAME,
                                             GENERATE_CONFIG_COMMAND_NAME)
+from syndicate.core.groups.tags import tags
 from syndicate.core.helper import (check_required_param,
                                    create_bundle_callback,
                                    handle_futures_progress_bar,
                                    resolve_path_callback, timeit,
-                                   verify_bundle_callback,
-                                   verify_meta_bundle_callback,
+                                   verify_bundle_callback, sync_lock,
                                    resolve_default_value,
                                    generate_default_bundle_name,
-                                   sync_lock)
+                                   resolve_and_verify_bundle_callback)
 from syndicate.core.project_state.project_state import (MODIFICATION_LOCK,
                                                         WARMUP_LOCK)
 from syndicate.core.project_state.status_processor import project_state_status
 from syndicate.core.project_state.sync_processor import sync_project_state
+from syndicate.core.constants import TEST_ACTION, BUILD_ACTION, \
+    DEPLOY_ACTION, UPDATE_ACTION, CLEAN_ACTION, SYNC_ACTION, \
+    STATUS_ACTION, WARMUP_ACTION, PROFILER_ACTION, ASSEMBLE_JAVA_MVN_ACTION, \
+    ASSEMBLE_PYTHON_ACTION, ASSEMBLE_NODE_ACTION, ASSEMBLE_ACTION, \
+    PACKAGE_META_ACTION, CREATE_DEPLOY_TARGET_BUCKET_ACTION, UPLOAD_ACTION, \
+    COPY_BUNDLE_ACTION
+
 
 INIT_COMMAND_NAME = 'init'
 SYNDICATE_PACKAGE_NAME = 'aws-syndicate'
@@ -94,40 +102,35 @@ def syndicate():
         sys.exit(1)
 
 
-@syndicate.command(name='test')
+@syndicate.command(name=TEST_ACTION)
 @click.option('--suite', type=click.Choice(['unittest', 'pytest', 'nose'],
                                            case_sensitive=False),
               default='unittest')
-@click.option('--test_folder_name', nargs=1, default='tests')
+@click.option('--test_folder_name', nargs=1, default='tests',
+              help='Directory in the project that contains tests to run. '
+                   'Default folder: tests')
 @click.option('--errors_allowed', is_flag=True)
-@timeit()
+@timeit(action_name=TEST_ACTION)
 def test(suite, test_folder_name, errors_allowed):
     """Discovers and runs tests inside python project configuration path."""
     click.echo('Running tests...')
     import subprocess
     from syndicate.core import CONFIG
     project_path = CONFIG.project_path
-
     test_folder = os.path.join(project_path, test_folder_name)
     if not os.path.exists(test_folder):
         click.echo(f'Tests not found, \'{test_folder_name}\' folder is missing'
                    f' in \'{project_path}\'.')
         return
-
     test_lib_command_mapping = {
         'unittest': f'{sys.executable} -m unittest discover {test_folder} -v',
         'pytest': 'pytest --no-header -v',
         'nose': 'nosetests --verbose'
     }
 
-    workdir = os.getcwd()
-
-    if test_folder != os.path.join(project_path, 'tests'):
-        os.chdir(test_folder)
     command = test_lib_command_mapping.get(suite)
-    result = subprocess.run(command.split())
+    result = subprocess.run(command.split(), cwd=project_path)
 
-    os.chdir(workdir)
     if not errors_allowed:
         if result.returncode != 0:
             click.echo('Some tests failed. Exiting.')
@@ -135,7 +138,7 @@ def test(suite, test_folder_name, errors_allowed):
     click.echo('Tests passed.')
 
 
-@syndicate.command(name='build')
+@syndicate.command(name=BUILD_ACTION)
 @click.option('--bundle_name', nargs=1,
               callback=generate_default_bundle_name,
               help='Name of the bundle to build. '
@@ -145,15 +148,16 @@ def test(suite, test_folder_name, errors_allowed):
 @click.option('--errors_allowed', is_flag=True, default=False,
               help='Flag to continue bundle building if some tests fail')
 @click.pass_context
-@timeit(action_name='build')
+@check_deploy_bucket_exists
+@timeit(action_name=BUILD_ACTION)
 def build(ctx, bundle_name, force_upload, errors_allowed):
     """
     Builds bundle of an application
     """
     if if_bundle_exist(bundle_name=bundle_name) and not force_upload:
-        click.echo('Bundle name \'{0}\' already exists '
-                   'in deploy bucket. Please use another bundle '
-                   'name or delete the bundle'.format(bundle_name))
+        click.echo(f'Bundle name \'{bundle_name}\' already exists '
+                   f'in deploy bucket. Please use another bundle '
+                   f'name or delete the bundle')
         return
     ctx.invoke(test, errors_allowed=errors_allowed)
     ctx.invoke(assemble, bundle_name=bundle_name)
@@ -161,7 +165,8 @@ def build(ctx, bundle_name, force_upload, errors_allowed):
     ctx.invoke(upload, bundle_name=bundle_name, force=force_upload)
 
 
-@syndicate.command(name='deploy')
+@syndicate.command(name=DEPLOY_ACTION)
+@sync_lock(lock_type=MODIFICATION_LOCK)
 @click.option('--deploy_name',
               callback=resolve_default_value,
               help='Name of the deploy. Default value: name of the project')
@@ -187,8 +192,8 @@ def build(ctx, bundle_name, force_upload, errors_allowed):
 @click.option('--replace_output', is_flag=True, default=False,
               help='Replaces the existing deploy output')
 @check_deploy_name_for_duplicates
-@sync_lock(lock_type=MODIFICATION_LOCK)
-@timeit(action_name='deploy')
+@check_deploy_bucket_exists
+@timeit(action_name=DEPLOY_ACTION)
 def deploy(deploy_name, bundle_name, deploy_only_types, deploy_only_resources,
            deploy_only_resources_path, excluded_resources,
            excluded_resources_path, excluded_types, continue_deploy,
@@ -196,8 +201,6 @@ def deploy(deploy_name, bundle_name, deploy_only_types, deploy_only_resources,
     """
     Deploys the application infrastructure
     """
-    sync_project_state()
-    from syndicate.core import PROJECT_STATE
     if deploy_only_resources_path and os.path.exists(
             deploy_only_resources_path):
         deploy_resources_list = json.load(open(deploy_only_resources_path))
@@ -225,11 +228,10 @@ def deploy(deploy_name, bundle_name, deploy_only_types, deploy_only_resources,
                                                      replace_output)
     click.echo('Backend resources were deployed{0}.'.format(
         '' if deploy_success else ' with errors. See deploy output file'))
-    PROJECT_STATE.release_lock(MODIFICATION_LOCK)
-    sync_project_state()
 
 
-@syndicate.command(name='update')
+@syndicate.command(name=UPDATE_ACTION)
+@sync_lock(lock_type=MODIFICATION_LOCK)
 @click.option('--bundle_name',
               callback=resolve_default_value,
               help='Name of the bundle to deploy. '
@@ -246,7 +248,8 @@ def deploy(deploy_name, bundle_name, deploy_only_types, deploy_only_resources,
                    'while deploy')
 @click.option('--replace_output', nargs=1, is_flag=True, default=False)
 @check_deploy_name_for_duplicates
-@timeit(action_name='update')
+@check_deploy_bucket_exists
+@timeit(action_name=UPDATE_ACTION)
 def update(bundle_name, deploy_name, replace_output,
            update_only_resources,
            update_only_resources_path,
@@ -281,8 +284,8 @@ def update(bundle_name, deploy_name, replace_output,
         click.echo('Something went wrong during resources update')
 
 
-@syndicate.command(name='clean')
-@timeit(action_name='clean')
+@syndicate.command(name=CLEAN_ACTION)
+@sync_lock(lock_type=MODIFICATION_LOCK)
 @click.option('--deploy_name', nargs=1, callback=resolve_default_value,
               help='Name of the deploy.')
 @click.option('--bundle_name', nargs=1, callback=resolve_default_value,
@@ -303,6 +306,7 @@ def update(bundle_name, deploy_name, replace_output,
               help='If specified provided types will be excluded')
 @click.option('--rollback', is_flag=True,
               help='Remove failed deploy resources')
+@timeit(action_name=CLEAN_ACTION)
 def clean(deploy_name, bundle_name, clean_only_types, clean_only_resources,
           clean_only_resources_path, clean_externals, excluded_resources,
           excluded_resources_path, excluded_types, rollback):
@@ -334,13 +338,12 @@ def clean(deploy_name, bundle_name, clean_only_types, clean_only_resources,
         excluded_resources = tuple(
             set(excluded_resources + tuple(excluded_resources_list)))
     if rollback:
-        remove_failed_deploy_resources(deploy_name=deploy_name,
-                                       bundle_name=bundle_name,
-                                       clean_only_resources=clean_only_resources,
-                                       clean_only_types=clean_only_types,
-                                       excluded_resources=excluded_resources,
-                                       excluded_types=excluded_types,
-                                       clean_externals=clean_externals)
+        remove_failed_deploy_resources(
+            deploy_name=deploy_name, bundle_name=bundle_name,
+            clean_only_resources=clean_only_resources,
+            clean_only_types=clean_only_types,
+            excluded_resources=excluded_resources,
+            excluded_types=excluded_types, clean_externals=clean_externals)
     else:
         remove_deployment_resources(deploy_name=deploy_name,
                                     bundle_name=bundle_name,
@@ -352,7 +355,8 @@ def clean(deploy_name, bundle_name, clean_only_types, clean_only_resources,
     click.echo('AWS resources were removed.')
 
 
-@syndicate.command(name='sync')
+@syndicate.command(name=SYNC_ACTION)
+@check_deploy_bucket_exists
 @timeit()
 def sync():
     """
@@ -362,11 +366,12 @@ def sync():
     return sync_project_state()
 
 
-@syndicate.command(name='status')
+@syndicate.command(name=STATUS_ACTION)
 @click.option('--events', 'category', flag_value='events',
               help='Show event logs of the project')
 @click.option('--resources', 'category', flag_value='resources',
               help='Show a summary of the project resources')
+@check_deploy_bucket_exists
 @timeit()
 def status(category):
     """
@@ -375,7 +380,8 @@ def status(category):
     click.echo(project_state_status(category))
 
 
-@syndicate.command(name='warmup')
+@syndicate.command(name=WARMUP_ACTION)
+@sync_lock(lock_type=WARMUP_LOCK)
 @click.option('--bundle_name', nargs=1, callback=resolve_default_value,
               help='Name of the bundle. Should be specified with deploy_name'
                    ' parameter.')
@@ -391,8 +397,8 @@ def status(category):
 @click.option('--header_name', nargs=1, help='Name of authentication header.')
 @click.option('--header_value', nargs=1, help='Name of authentication header '
                                               'value.')
-@sync_lock(lock_type=WARMUP_LOCK)
-@timeit(action_name='warmup')
+@check_deploy_bucket_exists
+@timeit(action_name=WARMUP_ACTION)
 def warmup(bundle_name, deploy_name, api_gw_id, stage_name, lambda_auth,
            header_name, header_value):
     """
@@ -434,11 +440,12 @@ def warmup(bundle_name, deploy_name, api_gw_id, stage_name, lambda_auth,
     click.echo('Application resources have been warmed up.')
 
 
-@syndicate.command(name='profiler')
+@syndicate.command(name=PROFILER_ACTION)
 @click.option('--bundle_name', nargs=1, callback=resolve_default_value)
 @click.option('--deploy_name', nargs=1, callback=resolve_default_value)
 @click.option('--from_date', nargs=1, type=str)
 @click.option('--to_date', nargs=1, type=str)
+@check_deploy_bucket_exists
 def profiler(bundle_name, deploy_name, from_date, to_date):
     """
     Displays application Lambda metrics
@@ -459,11 +466,12 @@ def profiler(bundle_name, deploy_name, from_date, to_date):
 # =============================================================================
 
 
-@syndicate.command(name='assemble_java_mvn')
+@syndicate.command(name=ASSEMBLE_JAVA_MVN_ACTION)
 @timeit()
 @click.option('--bundle_name', nargs=1, callback=generate_default_bundle_name)
 @click.option('--project_path', '-path', nargs=1,
               callback=resolve_path_callback)
+@timeit(action_name=ASSEMBLE_JAVA_MVN_ACTION)
 def assemble_java_mvn(bundle_name, project_path):
     """
     Builds Java lambdas
@@ -471,18 +479,19 @@ def assemble_java_mvn(bundle_name, project_path):
     :param project_path: path to project folder
     :return:
     """
-    click.echo('Command compile java project path: %s' % project_path)
+    click.echo(f'Command compile java project path: {project_path}')
     assemble_artifacts(bundle_name=bundle_name,
                        project_path=project_path,
                        runtime=RUNTIME_JAVA_8)
     click.echo('Java artifacts were prepared successfully.')
 
 
-@syndicate.command(name='assemble_python')
+@syndicate.command(name=ASSEMBLE_PYTHON_ACTION)
 @timeit()
 @click.option('--bundle_name', nargs=1, callback=generate_default_bundle_name)
 @click.option('--project_path', '-path', nargs=1,
               callback=resolve_path_callback)
+@timeit(action_name=ASSEMBLE_PYTHON_ACTION)
 def assemble_python(bundle_name, project_path):
     """
     Builds Python lambdas
@@ -490,18 +499,19 @@ def assemble_python(bundle_name, project_path):
     :param project_path: path to project folder
     :return:
     """
-    click.echo('Command assemble python: project_path: %s ' % project_path)
+    click.echo(f'Command assemble python: project_path: {project_path} ')
     assemble_artifacts(bundle_name=bundle_name,
                        project_path=project_path,
                        runtime=RUNTIME_PYTHON)
     click.echo('Python artifacts were prepared successfully.')
 
 
-@syndicate.command(name='assemble_node')
+@syndicate.command(name=ASSEMBLE_NODE_ACTION)
 @timeit()
 @click.option('--bundle_name', nargs=1, callback=generate_default_bundle_name)
 @click.option('--project_path', '-path', nargs=1,
               callback=resolve_path_callback)
+@timeit(action_name=ASSEMBLE_NODE_ACTION)
 def assemble_node(bundle_name, project_path):
     """
     Builds NodeJS lambdas
@@ -509,7 +519,7 @@ def assemble_node(bundle_name, project_path):
     :param project_path: path to project folder
     :return:
     """
-    click.echo('Command assemble node: project_path: %s ' % project_path)
+    click.echo(f'Command assemble node: project_path: {project_path} ')
     assemble_artifacts(bundle_name=bundle_name,
                        project_path=project_path,
                        runtime=RUNTIME_NODEJS)
@@ -523,10 +533,11 @@ RUNTIME_LANG_TO_BUILD_MAPPING = {
 }
 
 
-@syndicate.command(name='assemble')
-@timeit()
-@click.option('--bundle_name', nargs=1, callback=generate_default_bundle_name)
+@syndicate.command(name=ASSEMBLE_ACTION)
+@click.option('--bundle_name', callback=generate_default_bundle_name,
+              help='Bundle\'s name to build the lambdas in')
 @click.pass_context
+@timeit(action_name=ASSEMBLE_ACTION)
 def assemble(ctx, bundle_name):
     """
     Builds the application artifacts
@@ -535,7 +546,7 @@ def assemble(ctx, bundle_name):
         will be associated
     :return:
     """
-    click.echo('Building artifacts ...')
+    click.echo(f'Building artifacts, bundle: {bundle_name}')
     from syndicate.core import PROJECT_STATE
     build_mapping_dict = PROJECT_STATE.load_project_build_mapping()
     if build_mapping_dict:
@@ -545,14 +556,15 @@ def assemble(ctx, bundle_name):
                 ctx.invoke(func, bundle_name=bundle_name,
                            project_path=value)
             else:
-                click.echo('Build tool is not supported: %s' % key)
+                click.echo(f'Build tool is not supported: {key}')
     else:
         click.echo('Projects to be built are not found')
 
 
-@syndicate.command(name='package_meta')
-@timeit()
-@click.option('--bundle_name', nargs=1, callback=verify_bundle_callback)
+@syndicate.command(name=PACKAGE_META_ACTION)
+@click.option('--bundle_name', required=True, callback=verify_bundle_callback,
+              help='Bundle\'s name to package the current meta in')
+@timeit(action_name=PACKAGE_META_ACTION)
 def package_meta(bundle_name):
     """
     Generates metadata about the application infrastructure
@@ -560,13 +572,13 @@ def package_meta(bundle_name):
     :return:
     """
     from syndicate.core import CONFIG
-    click.echo('Package meta, bundle: %s' % bundle_name)
+    click.echo(f'Package meta, bundle: {bundle_name}')
     create_meta(project_path=CONFIG.project_path,
                 bundle_name=bundle_name)
     click.echo('Meta was configured successfully.')
 
 
-@syndicate.command(name='create_deploy_target_bucket')
+@syndicate.command(name=CREATE_DEPLOY_TARGET_BUCKET_ACTION)
 @timeit()
 def create_deploy_target_bucket():
     """
@@ -574,15 +586,20 @@ def create_deploy_target_bucket():
     :return:
     """
     from syndicate.core import CONFIG
-    click.echo('Create deploy target sdk: %s' % CONFIG.deploy_target_bucket)
+    click.echo(f'Create deploy target sdk: {CONFIG.deploy_target_bucket}')
     create_bundles_bucket()
     click.echo('Deploy target bucket was created successfully')
 
 
-@syndicate.command(name='upload')
-@click.option('--bundle_name', nargs=1, callback=verify_meta_bundle_callback)
-@click.option('--force', is_flag=True)
-@timeit(action_name='upload')
+@syndicate.command(name=UPLOAD_ACTION)
+@click.option('--bundle_name', callback=resolve_and_verify_bundle_callback,
+              help='Bundle name to which the build artifacts are gathered and '
+                   'later used for the deployment. NOTE: if not specified, '
+                   'the latest build will be uploaded')
+@click.option('--force', is_flag=True, help='Flag to override existing bundle '
+                                            'with the same name as provided')
+@check_deploy_bucket_exists
+@timeit(action_name=UPLOAD_ACTION)
 def upload(bundle_name, force=False):
     """
     Uploads bundle from local storage to AWS S3
@@ -591,7 +608,7 @@ def upload(bundle_name, force=False):
         already exists in an account
     :return:
     """
-    click.echo('Upload bundle: %s' % bundle_name)
+    click.echo(f'Upload bundle: {bundle_name}')
     if force:
         click.echo('Force upload')
 
@@ -601,7 +618,7 @@ def upload(bundle_name, force=False):
     click.echo('Bundle was uploaded successfully')
 
 
-@syndicate.command(name='copy_bundle')
+@syndicate.command(name=COPY_BUNDLE_ACTION)
 @click.option('--bundle_name', nargs=1, callback=create_bundle_callback)
 @click.option('--src_account_id', '-acc_id', nargs=1,
               callback=check_required_param)
@@ -630,11 +647,11 @@ def copy_bundle(ctx, bundle_name, src_account_id, src_bucket_region,
         already exists in a target account
     :return:
     """
-    click.echo('Copy bundle: %s' % bundle_name)
-    click.echo('Bundle name: %s' % bundle_name)
-    click.echo('Source account id: %s' % src_account_id)
-    click.echo('Source bucket region: %s' % src_bucket_region)
-    click.echo('Source bucket name: %s' % src_bucket_name)
+    click.echo(f'Copy bundle: {bundle_name}')
+    click.echo(f'Bundle name: {bundle_name}')
+    click.echo(f'Source account id: {src_account_id}')
+    click.echo(f'Source bucket region: {src_bucket_region}')
+    click.echo(f'Source bucket name: {src_bucket_name}')
     futures = load_bundle(bundle_name, src_account_id, src_bucket_region,
                           src_bucket_name, role_name)
     handle_futures_progress_bar(futures)
@@ -644,3 +661,4 @@ def copy_bundle(ctx, bundle_name, src_account_id, src_bucket_region,
 
 
 syndicate.add_command(generate)
+syndicate.add_command(tags)