+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Interface of command was improved, `access_key` and `secret_key` params made optional
 - Removed `account_id` param and provided ability to resolve it by STS or instance profile
 - `python_build_mapping`, `java_build_mapping`, `nodejs_build_mapping` was removed as they will be resolved using the `.syndicate` file
-- `generate lambda` command was improved, `common` module for python runtime is being generated with the first lambda
+- `generate lambda` command was improved, `common` and `tests` module for python runtime is generated with the first lambda. They contain basic common class all lambdas and simple initial test case.
 - Provided ability to add information about created lambda to `.syndicate` file
 - Added `test` command that allows to run tests inside your python project
 - Added `profiler` command that allows to display application Lambda metrics